% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacpd_mw.R
\name{lacpd_mw}
\alias{lacpd_mw}
\title{Locally adaptive change-point detection using the Mann-Whitney test}
\usage{
lacpd_mw(
  x,
  m = 1,
  k = c(2:5),
  blow = 0.1,
  bup = (1 - blow),
  adjust = FALSE,
  history = FALSE,
  eps = 1,
  alpha = 0.05,
  double = FALSE,
  ...
)
}
\arguments{
\item{x}{a numeric vector}

\item{m}{number of times to sub-sample}

\item{k}{single number or numeric vector proportional to the number of points on each side of the target point. See details}

\item{blow}{fraction of observations (0-1) at the beginning of the time-series not considered for change detection}

\item{bup}{similar to \code{blow}, but for the end of the time series. Default is 1-\code{blow}}

\item{adjust}{if \code{TRUE}, p-value will be adjusted by methods in \link[stats]{p.adjust}}

\item{history}{if \code{TRUE}, it maintains the step-wise results when \code{k} is a vector}

\item{eps}{an integer that will be used as a stopping rule. Once the difference between three different sliding windows in a row is less than eps, the algorithm stops.}

\item{alpha}{significance level}

\item{double}{logical. if TRUE it adjust p-values two times.}

\item{...}{arguments passed to \link[stats]{p.adjust}}
}
\value{
adaptive window: the automatic selected adaptive window.

change-point: the index of the most probable change-point in the time series

statistics: the Z statistics of LACPD

magnitude: the magnitude of change

p.value: the corresponding p.value


Attributes:

attr(,"zs"): retrieves the obtained Z statistics at the time-periods the function has looked for potential changes

attr(,"ps"): retrieves the obtained p-values at the time-periods the function has looked for potential changes

attr(,"mags"): retrieves the magnitude of change at the time-periods the function has looked for potential changes


if \code{history=TRUE} and \code{k} is a numeric vector, then the following attributes can also be retrieved.

attr(,"history"): a dataframe containing the results (cp, magnitude, Z, and p.value) based on different adaptive sliding windows which are permutations of the given \code{k}

attr(,"allzs"): a list which retrieves the obtained Z statistics at the time-periods the function has looked for potential changes based on different adaptive sliding windows which are permutations of the given \code{k}

attr(,"allmags"): a list which retrieves the magnitude of change at the time-periods the function has looked for potential changes based on different adaptive sliding windows which are permutations of the given \code{k}

attr(,"allps"): a list which retrieves the obtained p-values at the time-periods the function has looked for potential changes based on different adaptive sliding windows which are permutations of the given \code{k}

attr(,"zmat"): a matrix of z statistics for each individual sliding window

attr(,"pmat"): a matrix of p-values for each individual sliding window

attr(,"mmat"): a matrix of magnitudes for each individual sliding window
}
\description{
Locally adaptive change-point detection after accommodating the Mann-Whitney test in the LACPD procedure
}
\details{
This technique accommodates the \link[stats]{wilcox.test} in the LACPD procedure of Moradi et al. (2020) to look for potential change-points in the numerical vector \code{x}.

The LACPD procedure is based on an iteration of several sliding windows of different widths. In the tails of \code{x}, since there are not enough data points before/after the target points, the method uses sub-sampling, wherein it moderates the effect of sub-sampling by \code{m} iteration.

Assume the length of \code{x} is n. The argument \code{k} is used to set the number of data points in each sides of the target point when looking for change-points. \code{n}/\code{k}  is the number of points on each side we consider. For instance, if \code{n=300} and \code{k=10}, this means we consider \code{30} observations before and 30 after the target point when locally detecting changes.

If \code{adjust=TRUE}, methods susch as "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none" can be passed through, e.g. \code{method="BY"}.

If \code{k} is a vector of numbers, then the function returns a result based on adaptive sliding windows which is an average result obtained from different windows.
}
\examples{
x <- rnorm(100)
Z <- lacpd_mw(x,m=10,k=c(2:5),history = TRUE,adjust = TRUE,method="BY")
plot(Z,type="l",ylab = "Z",xlab = "time")
}
\references{
Moradi, M., Montesino-SanMartin, M., Ugarte, M. D., and Militino, A. F. (2021). Locally adaptive change-point detection with applications to remote sensing and land use changes.
}
\seealso{
\link[stats]{wilcox.test}, \link[stats]{p.adjust}
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
