% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lacpd.mw.R
\name{plot.lacpd.mw}
\alias{plot.lacpd.mw}
\title{Classes and methods for class lacpd.mw}
\usage{
\method{plot}{lacpd.mw}(x, ...)
}
\arguments{
\item{x}{an object of class lacpd.mw}
}
\description{
Classes and methods for class lacpd.mw
}
\details{
plot functions for class lacpd.mw
}
\seealso{
\link[LACPD]{lacpd_mw}
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
