% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lacpd.mw.R
\name{print.lacpd.mw}
\alias{print.lacpd.mw}
\title{Classes and methods for class lacpd.mw}
\usage{
\method{print}{lacpd.mw}(x, round = 5, ...)
}
\arguments{
\item{x}{an object of class lacpd.mw}
}
\description{
Classes and methods for class lacpd.mw
}
\details{
print function for class lacpd.mw
}
\seealso{
\link[LACPD]{lacpd_mw}
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
